requirements.txt
mockredispy==2.9.3
redis==2.10.5
wsgiref==0.1.2
